backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
python-dotenv==1.0.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.8.2
